Ports some Java changes to the C++ Ticl.
Specifically, moves the batching task out of the protocol handler and into
the invalidation client core.

Original change description:

Implements (un)marshalling of in-memory Ticl state to protocol buffers.

This CL defines protocol buffers to represent various parts of the Ticl
in-memory state and adds code to the Java Ticl to (un)marshall to and from
them. Additionally, as part of this change, it completes the transition of
ProtocolHandler to a passive class by moving the recurring task for batching
from ProtocolHandler to InvalidationClientCore.



git-svn-id: http://google-cache-invalidation-api.googlecode.com/svn/trunk/src@255 1cc9d426-c294-39be-ba72-c0199ca0f247